MODIFY: service 컬럼값 entity폴더로 분리
